update MoreInfo component as a wrapper
