--- packages/components/src/local/molecules/info-card/README.md
@@ -1,27 +1,0 @@
-# InfoCard documentation
-
-This is an example readme file for the 'info-card' component.
-
-## UI Example
-
-<!-- STORY -->
-
-## Code example
-
-<!-- SOURCE -->
-
-## Props
-
-<!-- PROPS -->
-
-## Usage
-
-### ES6 Import
-```js
-import InfoCard from './molecules/info-card'
-```
-
-### CommonJS
-
-```js
-const InfoCard = require('./molecules/info-card')
