--- packages/components/src/local/molecules/info-card/types/props.d.ts
@@ -1,4 +1,0 @@
-export default interface InfoCardProps {
-  content: string
-  title?: string
-};
