+++ packages/components/src/local/molecules/more-info/README.md
@@ -1,0 +1,27 @@
+# MoreInfo documentation
+
+This is an example readme file for the 'MoreInfo' component.
+
+## UI Example
+
+<!-- STORY -->
+
+## Code example
+
+<!-- SOURCE -->
+
+## Props
+
+<!-- PROPS -->
+
+## Usage
+
+### ES6 Import
+```js
+import MoreInfo from './molecules/more-info'
+```
+
+### CommonJS
+
+```js
+const MoreInfo = require('./molecules/more-info')
