+++ packages/components/src/local/molecules/more-info/styles.module.scss
@@ -17,8 +17,6 @@
   font-size: small;
 }
 
-.toggleDisplayMode {
-  font-size: smaller;
-  font-weight: 700;
-  cursor: pointer;
+.childNode {
+  cursor: help;
 }
