+++ packages/components/src/local/molecules/more-info/styles.module.scss.d.ts
@@ -1,10 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'infoCard': string;
-  'header': string;
+  'childNode': string;
   'content': string;
-  'toggleDisplayMode': string;
+  'header': string;
+  'infoCard': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
