+++ packages/components/src/local/molecules/more-info/types/props.d.ts
@@ -1,0 +1,6 @@
+import React from 'react'
+
+export default interface Props {
+  description: string
+  children: React.ReactNode
+};
