+++ packages/components/src/local/organisms/setting-platform-types/styles.module.scss.d.ts
@@ -3,13 +3,13 @@
 interface CssExports {
   'actions': string;
   'col': string;
+  'commoditie-header': string;
   'form-row': string;
   'mobile': string;
   'mobile-title': string;
   'row': string;
   'section': string;
   'states-holder': string;
-  'commoditie-header': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
